requirements.txt
python-dotenv
Django
djangorestframework
djangorestframework-simplejwt
psycopg[binary]>=3.2
pillow
drf-spectacular
django-cors-headers
gunicorn
uvicorn[standard]
# whitenoise
django-storages
boto3
